reorganized files into a components folder :file_folder:
